Move email resources to dedicated pkg
