+++ internal/auth0/email/resource_test.go
@@ -1,14 +1,11 @@
-package provider
+package email_test
 
 import (
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
-	"github.com/auth0/go-auth0/management"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/stretchr/testify/assert"
 
+	"github.com/auth0/terraform-provider-auth0/internal/provider"
 	"github.com/auth0/terraform-provider-auth0/internal/recorder"
 	"github.com/auth0/terraform-provider-auth0/internal/sweep"
 )
@@ -165,7 +162,7 @@ func TestAccEmail(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: TestFactories(httpRecorder),
+		ProviderFactories: provider.TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCreateSESEmailProvider,
@@ -283,41 +280,5 @@ func TestAccEmail(t *testing.T) {
 				),
 			},
 		},
-	})
-}
-
-func TestEmailProviderIsConfigured(t *testing.T) {
-	t.Run("it returns true if the provider is configured", func(t *testing.T) {
-		testHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path == "/api/v2/emails/provider" {
-				w.WriteHeader(http.StatusOK)
-				return
-			}
-			http.NotFound(w, r)
-		})
-		testServer := httptest.NewServer(testHandler)
-
-		api, err := management.New(testServer.URL, management.WithInsecure())
-		assert.NoError(t, err)
-
-		actual := emailProviderIsConfigured(api)
-		assert.True(t, actual)
-	})
-
-	t.Run("it returns false if the provider is not configured", func(t *testing.T) {
-		testHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path == "/api/v2/emails/provider" {
-				http.NotFound(w, r)
-				return
-			}
-			http.NotFound(w, r)
-		})
-		testServer := httptest.NewServer(testHandler)
-
-		api, err := management.New(testServer.URL, management.WithInsecure())
-		assert.NoError(t, err)
-
-		actual := emailProviderIsConfigured(api)
-		assert.False(t, actual)
 	})
 }
