+++ internal/provider/provider.go
@@ -17,6 +17,7 @@ import (
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/client"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/connection"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/customdomain"
+	"github.com/auth0/terraform-provider-auth0/internal/auth0/email"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/guardian"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/hook"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/logstream"
@@ -105,8 +106,8 @@ func New() *schema.Provider {
 			"auth0_hook":                       hook.NewResource(),
 			"auth0_prompt":                     prompt.NewResource(),
 			"auth0_prompt_custom_text":         prompt.NewCustomTextResource(),
-			"auth0_email":                      newEmail(),
-			"auth0_email_template":             newEmailTemplate(),
+			"auth0_email":                      email.NewResource(),
+			"auth0_email_template":             email.NewTemplateResource(),
 			"auth0_user":                       user.NewResource(),
 			"auth0_tenant":                     tenant.NewResource(),
 			"auth0_role":                       role.NewResource(),
